requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
gspread==3.6.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dotenv==0.13.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
